requirements.txt
openai>=1.0.0,<2.0.0
jsonschema>=4.0.0
click>=8.0.0
python-dotenv>=1.0.0
colorama>=0.4.0
